sensors: fix a potential sensor events may block issue
There's report issue that the sensor polling events may block at
NativeSensorManager::readEvents. The issue is happen when there's
exactly count sensor events in the input subsystem. At this case,
the POLLIN flag is not set on the following call of
sensor_poll_context_t::pollEvents and it will be stuck at
NativeSensorManager::readEvents until the next sensor events is
reported.

Change-Id: Id209e1db96585ac9dc39fcbef77897dff4e758cd